[Pdm2-9] Move RestrictedLayout to properly work with Domain-Based organization
