--- src/app/(Restricted)/index.tsx
@@ -1,6 +1,0 @@
-import RolesScreen from "./Roles/screens/RolesScreen";
-import React from "react";
-
-export default function index() {
-  return <RolesScreen />;
-}
